+++ packages/hook-server/src/server.ts
@@ -26,9 +26,9 @@ import { handlePromptSubmit } from "./handlers/prompt-submit";
 import { handleConfigChange } from "./handlers/config-change";
 import { StdoutApprovalChannel, AutoApproveApprovalChannel } from "./approvals/index";
 import type { ApprovalChannel } from "./approvals/index";
+export type { ApprovalChannel, ApprovalRequest } from "./approvals/index";
 
 export { AutoApproveApprovalChannel };
-export type { ApprovalChannel } from "./approvals/index";
 export { TelegramApprovalChannel } from "./approvals/index";
 export { WebhookApprovalChannel } from "./approvals/index";